requirements.txt
pycrypto==2.6.1
suds==0.4
